project finished

+++ contrat_project/src/main/java/uadb/location/contrat/dto/controller/problemeController/InfoProblemeWORelationsResponse.java
@@ -5,9 +5,9 @@ import uadb.location.contrat.model.probleme.Probleme;
 import java.time.LocalDateTime;
 
 public record InfoProblemeWORelationsResponse(
-        Integer id,
-        Integer contratId,
-        Integer signalePar,
+        Long id,
+        Long contratId,
+        Long signalePar,
         String description,
         Probleme.Type type,
         Probleme.Responsable responsable,
